+++ src/main/java/com/amaze/filemanager/fragments/PropertiesSheet.java
@@ -16,7 +16,7 @@ import android.widget.TextView;
 
 import com.amaze.filemanager.R;
 import com.amaze.filemanager.filesystem.BaseFile;
-import com.amaze.filemanager.utils.Futils;
+import com.amaze.filemanager.utils.files.Futils;
 import com.amaze.filemanager.utils.Utils;
 
 import java.io.File;
