+++ src/main/java/com/amaze/filemanager/services/asynctasks/LoadList.java
@@ -40,10 +40,10 @@ import com.amaze.filemanager.fragments.CloudSheetFragment;
 import com.amaze.filemanager.fragments.MainFragment;
 import com.amaze.filemanager.ui.LayoutElement;
 import com.amaze.filemanager.ui.icons.Icons;
-import com.amaze.filemanager.utils.CloudUtil;
-import com.amaze.filemanager.utils.CryptUtil;
+import com.amaze.filemanager.utils.cloud.CloudUtil;
+import com.amaze.filemanager.utils.files.CryptUtil;
 import com.amaze.filemanager.utils.DataUtils;
-import com.amaze.filemanager.utils.FileListSorter;
+import com.amaze.filemanager.utils.files.FileListSorter;
 import com.amaze.filemanager.utils.HistoryManager;
 import com.amaze.filemanager.utils.OTGUtil;
 import com.amaze.filemanager.utils.OpenMode;
