+++ src/main/java/com/amaze/filemanager/ui/views/preference/CheckBox.java
@@ -1,4 +1,4 @@
-package com.amaze.filemanager.ui.views;
+package com.amaze.filemanager.ui.views.preference;
 
 import android.content.Context;
 import android.preference.SwitchPreference;
