+++ src/main/java/com/amaze/filemanager/ui/views/preference/EditIntPreference.java
@@ -1,8 +1,7 @@
-package com.amaze.filemanager.ui.views;
+package com.amaze.filemanager.ui.views.preference;
 
 import android.content.Context;
 import android.preference.EditTextPreference;
-import android.util.AttributeSet;
 
 /**
  * Created by rpiotaix on 17/11/16.
