+++ src/main/java/com/amaze/filemanager/utils/files/FileListSorter.java
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.amaze.filemanager.utils;
+package com.amaze.filemanager.utils.files;
 
 import com.amaze.filemanager.ui.LayoutElement;
