+++ src/main/res/xml/advancedsearch_prefs.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
-    <com.amaze.filemanager.ui.views.CheckBox
+    <com.amaze.filemanager.ui.views.preference.CheckBox
         android:key="regex"
         android:summary="@string/regular_expression_summary"
         android:title="@string/regular_expression"
         />
-    <com.amaze.filemanager.ui.views.CheckBox
+    <com.amaze.filemanager.ui.views.preference.CheckBox
         android:key="matches"
         android:summary="@string/match_content_summary"
         android:title="@string/match_content"
