+++ src/main/res/xml/preferences.xml
@@ -27,26 +27,26 @@
             android:summary="@string/color_summary"
             android:title="@string/color_title"/>
 
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="true"
             android:key="circularimages"
             android:summary="@string/circularimages"
             android:title="@string/circularicons"/>
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="false"
             android:enabled="true"
             android:key="plus_pic"
             android:summary="@string/gplus_pic_summary"
             android:title="@string/gplus_pic_title"/>
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="true"
             android:key="showDividers"
             android:title="@string/showDividers"/>
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="true"
             android:key="showHeaders"
             android:title="@string/showHeaders"/>
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="true"
             android:key="intelliHideToolbar"
             android:summary="@string/intellihide_toolbar_summary"
@@ -55,7 +55,7 @@
     </PreferenceCategory>
 
     <PreferenceCategory android:title="SideBar">
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="true"
             android:key="sidebar_folders_enable"
             android:title="@string/showfolders_pref_title"/>
@@ -64,7 +64,7 @@
             android:key="sidebar_folders"
             android:title="@string/showfolders_pref"/>
 
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="true"
             android:key="sidebar_quickaccess_enable"
             android:title="@string/showquickaccess_pref_title"/>
@@ -75,31 +75,31 @@
     </PreferenceCategory>
 
     <PreferenceCategory android:title="@string/general">
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="true"
             android:key="savepaths"
             android:summary="@string/savepathsummary"
             android:title="@string/savepaths"/>
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="true"
             android:key="showThumbs"
             android:summary="@string/thumbSummary"
             android:title="@string/thumb"/>
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="false"
             android:key="showHidden"
             android:title="@string/hidden"/>
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="true"
             android:key="showLastModified"
             android:summary="@string/lastModifiedSummary"
             android:title="@string/lastModifiedPref"/>
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="false"
             android:key="showFileSize"
             android:summary="@string/sizePrefSummary"
             android:title="@string/sizePref"/>
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="false"
             android:key="goBack_checkbox"
             android:summary="@string/back_summary"
@@ -130,13 +130,13 @@
             android:title="@string/crypt_pref_master_password_title">
         </EditTextPreference>
 
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:key="crypt_fingerprint"
             android:summary="@string/crypt_pref_fingerprint_summary"
             android:title="@string/crypt_pref_fingerprint_title"
             android:enabled="false"
             >
-        </com.amaze.filemanager.ui.views.CheckBox>
+        </com.amaze.filemanager.ui.views.preference.CheckBox>
     </PreferenceCategory>
 
     <PreferenceCategory android:title="@string/miscellaneous">
@@ -145,12 +145,12 @@
             android:key="advancedsearch"
             android:title="@string/advanced_search"/>
 
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="false"
             android:key="showPermissions"
             android:summary="@string/rootPrefSummary"
             android:title="@string/rootPref"/>
-        <com.amaze.filemanager.ui.views.CheckBox
+        <com.amaze.filemanager.ui.views.preference.CheckBox
             android:defaultValue="false"
             android:key="rootmode"
             android:summary="@string/rootmodesummary"
